[refactor] ディレクトリを整理
