+++ api-schema/src/types/roomState.ts
@@ -3,5 +3,6 @@ import { RoomMember } from "@api-schema/types/member";
 export type RoomState = {
   roomId: number;
   members: RoomMember[];
+  //カメラ側のstreamIdも合った方がいいかも
   focusStreamId: string | null;
 };
