+++ client/src/components/page/space/memberBlock/MemberItem.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 import { Avatar, Text, VStack } from "@chakra-ui/react";
-import { AvatarReactionBadge } from "../../common/AvatarReactionBadge";
-import { MemberStatus } from "../../../hooks/useLegacySyncMemberStatus";
-import { Member } from "../../../hooks/useRoom";
+import { AvatarReactionBadge } from "../../../common/AvatarReactionBadge";
+import { MemberStatus } from "../../../../hooks/useLegacySyncMemberStatus";
+import { Member } from "../../../../hooks/useRoom";
 
 //別のとこに書いたほうがよさそう
