+++ client/src/components/page/space/timetableBlock/TimetableBlock.tsx
