+++ client/src/hooks/useRoom.ts
@@ -2,7 +2,7 @@ import client from "../utils/api-client.factory";
 
 import useSWR from "swr";
 import { useMemo } from "react";
-import { UserType } from "../components/page/space/MemberItem";
+import { UserType } from "../components/page/space/memberBlock/MemberItem";
 import { UserResponse } from "@api-schema/api/@types";
 
 const roomFetcher = async (key: string, roomId: number) => {
