+++ client/src/states/useSkyway.ts
@@ -1,0 +1,15 @@
+import { atom } from "recoil";
+import { SkywayCredentialsModel } from "@api-schema/api/@types";
+
+// const videotelephonyCredentialState = atom<SkywayCredentialsModel | null>({
+//   key: "useSkywayCredential-videotelephonyCredentialState",
+//   default: null,
+// });
+//
+// const screenShareCredentialState = atom<SkywayCredentialsModel | null>({
+//   key: "useSkywayCredential-screenShareCredentialState",
+//   default: null,
+// });
+
+//setCredentialでcredentialをストアし接続
+//
